Fix Open Graph image route rendering in Next 16
